fixes and separation
