+++ pyproject.toml
@@ -130,6 +130,9 @@ ignore = [
 [tool.ruff.lint.flake8-builtins]
 builtins-ignorelist = ["input", "id", "bytes"]
 
+[tool.ruff.lint.flake8-pytest-style]
+fixture-parentheses = true
+
 [tool.ruff.lint.pydocstyle]
 convention = "numpy"
