+++ toolkit/risk_networks/config.py
@@ -7,6 +7,9 @@ from typing import TypedDict
 
 from toolkit.helpers.constants import CACHE_PATH
 
+SIMILARITY_THRESHOLD_MIN = 0.0
+SIMILARITY_THRESHOLD_MAX = 1.0
+
 list_sep = ";"
 max_rows_to_show = 1000
 entity_label = "ENTITY"
