--- toolkit/risk_networks/constants.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation. All rights reserved.
-# Licensed under the MIT license. See LICENSE file in the project.
-#
-
-SIMILARITY_THRESHOLD_MIN = 0.0
-SIMILARITY_THRESHOLD_MAX = 1.0
